refactor: move assistant logic into dedicated folder
