+++ listeners/assistant/__init__.py
@@ -1,0 +1,3 @@
+from .assistant import assistant
+
+__all__ = ["assistant"]
